add pool tests
