+++ x/concentrated-liquidity/pool_test.go
@@ -3,8 +3,10 @@ package concentrated_liquidity_test
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	cl "github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity"
 	"github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity/internal/math"
 	"github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity/types"
+	swaproutertypes "github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
 )
 
 func (s *KeeperTestSuite) TestCalcAndSwapOutAmtGivenIn() {
@@ -1128,4 +1130,30 @@ func (s *KeeperTestSuite) TestPoolExists() {
 
 	// ensure that this returns false
 	s.Require().False(poolExists)
+}
+
+func (s *KeeperTestSuite) TestConvertConcentratedToPoolInterface() {
+	s.SetupTest()
+
+	// Create default CL pool
+	concentratedPool := s.PrepareConcentratedPool()
+
+	// Ensure no error occurs when converting to PoolInterface
+	_, err := cl.ConvertConcentratedToPoolInterface(concentratedPool)
+	s.Require().NoError(err)
+}
+
+func (s *KeeperTestSuite) TestConvertPoolInterfaceToConcentrated() {
+	s.SetupTest()
+
+	// Create default CL pool
+	concentratedPool := s.PrepareConcentratedPool()
+
+	// Make it a poolI
+	poolI, ok := concentratedPool.(swaproutertypes.PoolI)
+	s.Require().True(ok)
+
+	// Ensure no error occurs when converting to PoolInterface
+	_, err := cl.ConvertPoolInterfaceToConcentrated(poolI)
+	s.Require().NoError(err)
 }
